Move the federated SAML UI to /ee
